[quant][refactor] Moving the 'quantization/fx/utils.py' to 'quantization/utils.py'
This file has more "generic" utilities, and if we try to use it from the parent, it will cause circular dependency error.

[ghstack-poisoned]
