+++ torch/quantization/__init__.py
@@ -6,7 +6,6 @@ from .fuse_modules import fuse_modules
 from .stubs import *
 from .quant_type import *
 from .quantize_jit import *
-# from .quantize_fx import *
 from .quantization_mappings import *
 from .fuser_method_mappings import *
